+++ lit-api-server/src/core/core_features.rs
@@ -2,10 +2,10 @@ use crate::accounts::can_execute_action;
 use crate::actions::client::ClientBuilder;
 use crate::actions::client::models::DenoExecutionEnv;
 use crate::actions::grpc::GrpcClientPool;
-use crate::core::api_status::ApiStatus;
+use crate::core::v1::helpers::api_status::ApiStatus;
 use crate::core::v1::models::request::LitActionRequest;
 use crate::core::v1::models::response::LitActionResponse;
-use crate::utils::parse_to_hash::ipfs_cid_to_u256;
+use crate::utils::parse_with_hash::ipfs_cid_to_u256;
 use ipfs_hasher::IpfsHasher;
 use moka::future::Cache;
 use rocket::serde::json::Json;
