+++ lit-api-server/src/utils/mod.rs
@@ -1,4 +1,4 @@
-pub mod parse_to_hash;
+pub mod parse_with_hash;
 use anyhow::Result;
 use ethers::types::{H160, U256};
 use ethers::utils::keccak256;
